pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "real-deal"
version = "0.1.0"
description = "Ontario real-estate deal scanner for cash-flowing properties"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "RealDeal" }]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "duckdb>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
real-deal = "real_deal.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
